requirements.txt
opencv-python~=4.9.0.80
jdeskew~=0.2.2
vietocr~=0.3.13
pillow~=10.2.0
numpy~=1.26.4
pandas~=2.2.1
PyYAML~=6.0.1
requests~=2.31.0